+++ code/spikes/publish.001/src/ui.react/Markdown.Outline/Markdown.Outline.RootSection.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { Color, COLORS, css, t, rx, FC } from '../common.mjs';
+import { Color, COLORS, css, t, rx, FC } from './common.mjs';
 
 export type MarkdownOutlineRootSectionProps = {
   index: number;
