+++ code/spikes/publish.001/src/ui.react/Markdown.Outline/common.mts
@@ -1,0 +1,2 @@
+export * from '../common.mjs';
+export { MarkdownUtil } from '../Markdown/Markdown.Util.mjs';
