+++ code/spikes/publish.001/src/ui.react/Markdown/index.mts
@@ -1,3 +1,2 @@
 export * from './Markdown';
-export * from './Markdown.Outline';
 export { MarkdownUtil } from './Markdown.Util.mjs';
